Acquire locked file assertion from UIProcess https://bugs.webkit.org/show_bug.cgi?id=271989 rdar://125037124
Reviewed by NOBODY (OOPS!).

We currently acquire a locked file assertion eagerly in the NetworkProcess when a SQLite transaction
starts. We would like to avoid doing this eagerly because that assertion (a FinishTaskInterruptible
assertion) prevents the system from sleeping, which is causing a power regression due to recent
changes in the OS surrounding the way background assertions are taken. We would also like to stop
taking the assertion from within NetworkProcess since we will likely not support taking assertions
directly from process extensions in the future.

To fix this, we defer taking any assertions unless we are about to suspend with an active SQLite
transaction (in `NetworkProcess::prepareToSuspend`). If there is an active transaction, then we
notify the UIProcess about that in the PrepareToSuspend IPC response and allow the UIProcess to
acquire the assertion synchronously. We add a `NetworkProcessProxy::releaseLockedFileAssertion` IPC
to drop the assertion when the last transaction completes.

* Source/WebKit/NetworkProcess/NetworkProcess.cpp:
(WebKit::NetworkProcess::NetworkProcess):
(WebKit::NetworkProcess::processWillSuspendImminentlyForTestingSync):
(WebKit::NetworkProcess::prepareToSuspend):
(WebKit::NetworkProcess::didFinishLastTransaction):
(WebKit::NetworkProcess::setIsHoldingLockedFiles): Deleted.
* Source/WebKit/NetworkProcess/NetworkProcess.h:
* Source/WebKit/NetworkProcess/NetworkProcess.messages.in:
* Source/WebKit/NetworkProcess/cocoa/NetworkProcessCocoa.mm:
(WebKit::NetworkProcess::acquireLockedFileActivity): Deleted.
(WebKit::NetworkProcess::invalidateFileActivity): Deleted.
(WebKit::NetworkProcess::hasAcquiredFileActivity const): Deleted.
* Source/WebKit/UIProcess/Network/NetworkProcessProxy.cpp:
(WebKit::NetworkProcessProxy::releaseLockedFileAssertion):
(WebKit::NetworkProcessProxy::sendPrepareToSuspend):
* Source/WebKit/UIProcess/Network/NetworkProcessProxy.h:
* Source/WebKit/UIProcess/Network/NetworkProcessProxy.messages.in: