Web Inspector: Dark Mode: Scope bars hovered and selected to match light mode https://bugs.webkit.org/show_bug.cgi?id=271988
Reviewed by NOBODY (OOPS!).

Purpose of pr is to darken text color when a scope bar is selected or hovered in dark mode (@media (prefers-color-scheme: dark)) to match when light mode is selected or hovered.
Create a variable --scope-bar-text-color-active and set to hsl(0, 0%, 15%) in ScopeBar.css.
When navigation bar is selected or hovered, set color to --scope-bar-text-color-active in RadioNavigationItem.css.
When sidebar is selected or hovered, set color to --scope-bar-text-color-active in GeneralStyleDetailsSidebarPanel.css.
When scope bar is selected or hovered, set color to --scope-bar-text-color-active in ScopeBar.css.

* Source/WebInspectorUI/UserInterface/Views/GeneralStyleDetailsSidebarPanel.css:
(@media (prefers-color-scheme: dark) .sidebar > .panel.details.css-style > .content ~ .options-container > .toggle:is(.selected,:hover)):
* Source/WebInspectorUI/UserInterface/Views/ScopeBar.css:
(@media (prefers-color-scheme: dark) .scope-bar > li:is(.selected, :hover)):
* Source/WebInspectorUI/UserInterface/Views/Variables.css:
(:root):