[Cocoa] Provide hints to CFNetwork for network traffic class https://bugs.webkit.org/show_bug.cgi?id=271920 <rdar://problem/125637205>
Reviewed by Jer Noble and Eric Carlson.

In Bug 254072 we started setting the traffic class for network activity in WebRTC
operations. We should take the same action when we know we are making high (or low)
priority requests to the network system so it can do a better job of routing packets
and organizing connections.

* Source/WebKit/NetworkProcess/cocoa/NetworkDataTaskCocoa.mm:
(WebKit::adjustNetworkServiceTypeIfNecessary): Added helper.
(WebKit::NetworkDataTaskCocoa::NetworkDataTaskCocoa):

Canonical link: https://commits.webkit.org/276901@main