gtk_widget_snapshot_child: assertion '_gtk_widget_get_parent (child) == widget' failed https://bugs.webkit.org/show_bug.cgi?id=271963
Reviewed by Carlos Garcia Campos.

Using a weak pointer to track the lifetime of the child dialog is
insufficient because we'll try to snapshot it after it has been
unparented but before it has been fully destroyed. It's guaranteed to be
unparented before dispose, so it's both necessary and sufficient to
watch notify::parent instead.

webkitWebViewBaseRemoveDialog is no longer needed since it becomes just
a one-liner.

* Source/WebKit/UIProcess/API/gtk/WebKitWebViewBase.cpp:
(webkitWebViewBaseAddDialog):
(webkitWebViewBaseDispose):
(webkitWebViewBaseRemoveDialog): Deleted.

Canonical link: https://commits.webkit.org/276950@main