AX: Mac layout tests shouldn't access attributes that aren't exposed for that element https://bugs.webkit.org/show_bug.cgi?id=271674 rdar://problem/125373694
Reviewed by Andres Gonzalez.

Checks accessibilityAttributeNames before calling accessibilityAttributeValue in
AccessibilityUIElementMac. Returns nil if querying for an attribute that isn't
found.

That led to finding lots of examples of attributes that ought to be exposed
in accessibilityAttributeNames but weren't now.

In this patch, I attempted to fix the ones that have a clear, straightforward
fix, and then added an exception for anything that wasn't straightforward.

The idea is to follow up until all of the exceptions are removed.

* LayoutTests/accessibility/lists-expected.txt: Removed.
* LayoutTests/accessibility/mac/aria-pressed-button-attributes-expected.txt:
* LayoutTests/accessibility/mac/aria-pressed-button-attributes.html:
* LayoutTests/accessibility/mac/aria-tree-with-presentation-role.html:
* LayoutTests/accessibility/mac/bounds-for-range-expected.txt:
* LayoutTests/accessibility/mac/internal-link-anchors-expected.txt:
* LayoutTests/accessibility/table-with-rules-expected.txt:
* LayoutTests/platform/mac/accessibility/lists-expected.txt:
* Source/WebCore/accessibility/AccessibilityObject.cpp:
(WebCore::AccessibilityObject::supportsExpanded const):
* Source/WebCore/accessibility/mac/WebAccessibilityObjectWrapperMac.mm:
(-[WebAccessibilityObjectWrapper _additionalAccessibilityAttributeNames:]):
(-[WebAccessibilityObjectWrapper ALLOW_DEPRECATED_IMPLEMENTATIONS_END]):
* Tools/WebKitTestRunner/InjectedBundle/mac/AccessibilityUIElementMac.mm:
(WTR::attributeValue):

Canonical link: https://commits.webkit.org/276955@main