REGRESSION: fix compilation if !ENABLE(VIDEO)
https://bugs.webkit.org/show_bug.cgi?id=271881

Reviewed by Mike Wyrzykowski.

[WebCore]: FullscreenManager: add missing include

HTMLMediaElement normally includes HTMLMediaElementEnums,
but only if VIDEO is enabled.

[WebCore]: FullscreenManager: fix undefined reference to HTMLMediaElement

HTMLMediaElement is not compiled if VIDEO is not enabled.

[WebCore]: css: fix fix undefined reference to HTMLMediaElement

HTMLMediaElement is not compiled if VIDEO is not enabled.

[WebCore]: InspectorDOMAgent: fix fix undefined reference to HTMLMediaElement

HTMLMediaElement is not compiled if VIDEO is not enabled.

[WebKit]: WebGPU: drop duplicate directive in RemoteDevice

The #if PLATFORM(COCOA) && ENABLE(VIDEO) directive is already defined
outside the function scope, so drop the inner ones.

[WebCore]: GPUDevice: fix undefined reference to HTMLMediaElement

HTMLMediaElement is not compiled if VIDEO is not enabled.

[WebKit]: WebFullScreenManagerProxy: add missing include

HTMLMediaElement normally includes HTMLMediaElementEnums,
but only if VIDEO is enabled.

[WebKit]: WebFullScreenManager: add missing include

HTMLMediaElement normally includes HTMLMediaElementEnums,
but only if VIDEO is enabled.

[WebKit]: InjectedBundlePageFullScreenClient: add missing include

HTMLMediaElement normally includes HTMLMediaElementEnums,
but only if VIDEO is enabled.

[WebKit]: WebCoreArgumentCoders.serialization.in: fix undefined reference

CaptionUserPreferencesDisplayMode and SerializedPlatformDataCueValue
are only compiled if VIDEO is enabled.

Signed-off-by: Thomas Devoogdt <thomas@devoogdt.com>
Canonical link: https://commits.webkit.org/276977@main