Validate key paths of animations in an animation group sent over IPC https://bugs.webkit.org/show_bug.cgi?id=271476 rdar://125216664
Reviewed by Simon Fraser.

Assert that the key path of animations created within an animation group is valid
when receiving them from an IPC connection in the same spirit as 259066@main.

Additionally, we add a `Validator` for the `PlatformCAAnimationRemoteProperties::animationType`
member such that it calls `PlatformCAAnimation::isValidKeyPath()` during the decoding
process and trigger a `MESSAGE_CHECK` if the key is not valid. It is important to do this for
`animationType` and not `keyPath` since we need both members to have been decoded first.

* Source/WebKit/WebProcess/WebPage/RemoteLayerTree/PlatformCAAnimationRemote.mm:
(WebKit::createAnimation):
* Source/WebKit/WebProcess/WebPage/RemoteLayerTree/PlatformCAAnimationRemoteProperties.serialization.in:

Canonical link: https://commits.webkit.org/272448.858@safari-7618-branch