Any early attempt to load the background page should return early https://bugs.webkit.org/show_bug.cgi?id=272120 rdar://125877776
Reviewed by Brian Weinstein.

We return early for premature background page loads. But we're not doing this after toggling the enabled
state of the extension. To fix this, we should set `m_safeToLoadBackgroundContent` back to false when
the extension unloads, since we use this variable to check if we should return early in
`loadBackgroundWebViewIfNeeded`.

* Source/WebKit/UIProcess/Extensions/Cocoa/WebExtensionContextCocoa.mm:
(WebKit::WebExtensionContext::unload):

Canonical link: https://commits.webkit.org/277041@main