REGRESSION (276827@main): [ MacOS wk1 Debug ] tables/mozilla/bugs/bug4576.html is a flaky crash (272128) https://bugs.webkit.org/show_bug.cgi?id=272128 rdar://125879096
Reviewed by Simon Fraser.

This assertion happens with DumpRenderTree on EWS only. It looks like there is
an animated image which is being decoded. But the animation of this image is reset
because a new page is loaded in the same WebProcess. When finishing the decoding,
the decoding thread returns the decoded frame to the BitmapImageSource which
sends it to ImageFrameAnimator. The ImageFrameAnimator sees that the animation
has not started yet on this page so the assertion fires.

An speculative fix is to change the assertion to an if-statement and make
ImageFrameAnimator returns false to BitmapImageSource in this case.

* LayoutTests/platform/mac-wk1/TestExpectations:
* Source/WebCore/platform/graphics/ImageFrameAnimator.cpp:
(WebCore::ImageFrameAnimator::imageFrameDecodeAtIndexHasFinished):

Canonical link: https://commits.webkit.org/277068@main