[WPE] Fix crash on ThreadedCompositor::renderLayerTree() during video playlist transitions https://bugs.webkit.org/show_bug.cgi?id=272150
Reviewed by Miguel Gomez.

Unnoticed default usage of dmabuf was leading to missing checks in the dmabuf
proxy. This caused issues when activating a proxy in a TextureMapperLayer,
removing any reference that another layer might have had to the previous proxy.
Additionally, when invalidating, the reference to the proxy held by the TextureMapperLayer
was also removed.

This prevents an invalid memory access identified in the
CoordinatedGraphicsScene::paintToCurrentGLContext().

* Source/WebCore/platform/graphics/texmap/TextureMapperPlatformLayerProxyDMABuf.cpp:
(WebCore::TextureMapperPlatformLayerProxyDMABuf::activateOnCompositingThread):
(WebCore::TextureMapperPlatformLayerProxyDMABuf::invalidate):

Canonical link: https://commits.webkit.org/277071@main